+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Search by WHO identifier
 - Sync progress loader and datetime
 - ANC details activity
+- Enroll into ANC and add Encounter, Observation, Episode of Care, Condition
 
 ### Fixed